requirements.txt
llama_index==0.11.20
llama-index-llms-anthropic==0.3.8
llama-index-llms-ollama==0.3.3
llama-index-llms-openai==0.2.10
llama-index-llms-openai-like==0.2.0
llama-index-llms-together==0.2.0
llama-index-utils-workflow==0.2.1
python-dotenv==1.0.1
shiny==1.1.0
pymongo==4.10.1
